+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file(see http://k
 - threadpool: user can retrieve the threads count.
 - fs: user can provide cwd in absolute method.
 - fs: add append method to append data to file.
+- fs: remove suffix from basename if user provided.
 - sys: retrieve current executable path.
 - sys: retrieve current process id.
 - sys: kill a specific process.